Deposit Account Schema Support
Source of truth: http://api.rebit.org.in/FISchema/deposit

A complete definition of this schema helps us use /FI/request
and /FI/fetch APIs and test them end-to-end.

The highlight of this implementation is that we have a complete
definition of all sorts of transactions, account holder's profile,
bank details and such modeled.